+++ packages/popup/src/locales/uk.js
